Finished implemented HRSC support

+++ include/genericls/UsgsAstroLsStateData.h
@@ -86,8 +86,8 @@ class UsgsAstroLsStateData
    int          m_AberrFlag;                      // 8
    int          m_AtmRefFlag;                     // 9
    std::vector<int> m_IntTimeLines;
-   std::vector<int> m_IntTimeStartTimes;
-   std::vector<int> m_IntTimes;
+   std::vector<double> m_IntTimeStartTimes;
+   std::vector<double> m_IntTimes;
    double       m_StartingEphemerisTime;          // 11
    double       m_CenterEphemerisTime;            // 12
    double       m_DetectorSampleSumming;          // 13
